@@ -1,22 +1,56 @@
 {
-  "name": "as-table",
-  "version": "1.0.36",
+  "_args": [
+    [
+      "as-table@1.0.52",
+      "/Users/Rothacker/Desktop/workspace/trading-bot"
+    ]
+  ],
+  "_from": "as-table@1.0.52",
+  "_id": "as-table@1.0.52",
+  "_inBundle": false,
+  "_integrity": "sha512-Lw3OjVL9BDGSrU6RoK5aTm79SnrDY0Zvucxc+oOTqZrC2Sw575IhkOTPtvpzviliL/WJaCSWzJdKFdYeI+8wFg==",
+  "_location": "/as-table",
+  "_phantomChildren": {},
+  "_requested": {
+    "type": "version",
+    "registry": true,
+    "raw": "as-table@1.0.52",
+    "name": "as-table",
+    "escapedName": "as-table",
+    "rawSpec": "1.0.52",
+    "saveSpec": null,
+    "fetchSpec": "1.0.52"
+  },
+  "_requiredBy": [
+    "/",
+    "/stacktracey"
+  ],
+  "_resolved": "https://registry.npmjs.org/as-table/-/as-table-1.0.52.tgz",
+  "_spec": "1.0.52",
+  "_where": "/Users/Rothacker/Desktop/workspace/trading-bot",
+  "author": {
+    "name": "Vitaly Gordon",
+    "email": "rocket.mind@gmail.com"
+  },
+  "bugs": {
+    "url": "https://github.com/xpl/as-table/issues"
+  },
+  "dependencies": {
+    "printable-characters": "^1.0.42"
+  },
   "description": "A simple function that print objects / arrays as ASCII tables. Handles ANSI styling and weird 💩 Unicode emoji symbols – they won't break the layout.",
-  "main": "build/as-table.js",
-  "typings": "./as-table.d.ts",
-  "scripts": {
-    "lint": "eslint as-table.js",
-    "lint-test": "eslint test.js",
-    "babel": "babel as-table.js --source-maps inline --out-file ./build/as-table.js",
-    "build": "npm run lint && npm run lint-test && npm run babel",
-    "coveralls": "nyc report --reporter=text-lcov | coveralls",
-    "test": "npm run build && env AS_TABLE_TEST_FILE='./build/as-table' nyc --reporter=html --reporter=text mocha --reporter spec",
-    "autotest": "env AS_TABLE_TEST_FILE='./as-table' mocha --reporter spec --watch"
+  "devDependencies": {
+    "ansicolor": "^1.1.81",
+    "babel-cli": "^6.26.0",
+    "babel-plugin-transform-es2015-destructuring": "^6.23.0",
+    "babel-plugin-transform-es2015-spread": "^6.22.0",
+    "coveralls": "^3.0.3",
+    "eslint": "^4.19.1",
+    "istanbul": "^0.4.5",
+    "mocha": "^6.0.2",
+    "nyc": "^13.3.0"
   },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/xpl/as-table.git"
-  },
+  "homepage": "https://github.com/xpl/as-table",
   "keywords": [
     "ASCII",
     "table",
@@ -44,24 +78,22 @@
     "text",
     "layout"
   ],
-  "author": "Vitaly Gordon <rocket.mind@gmail.com>",
   "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/xpl/as-table/issues"
+  "main": "build/as-table.js",
+  "name": "as-table",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/xpl/as-table.git"
   },
-  "homepage": "https://github.com/xpl/as-table",
-  "devDependencies": {
-    "ansicolor": "^1.1.71",
-    "babel-cli": "^6.26.0",
-    "babel-plugin-transform-es2015-destructuring": "^6.23.0",
-    "babel-plugin-transform-es2015-spread": "^6.22.0",
-    "coveralls": "^2.13.3",
-    "eslint": "^4.8.0",
-    "istanbul": "^0.4.5",
-    "mocha": "^3.5.3",
-    "nyc": "^11.2.1"
+  "scripts": {
+    "autotest": "env AS_TABLE_TEST_FILE='./as-table' mocha --reporter spec --watch",
+    "babel": "babel as-table.js --source-maps inline --out-file ./build/as-table.js",
+    "build": "npm run lint && npm run lint-test && npm run babel",
+    "coveralls": "nyc report --reporter=text-lcov | coveralls",
+    "lint": "eslint as-table.js",
+    "lint-test": "eslint test.js",
+    "test": "npm run build && env AS_TABLE_TEST_FILE='./build/as-table' nyc --reporter=html --reporter=text mocha --reporter spec"
   },
-  "dependencies": {
-    "printable-characters": "^1.0.37"
-  }
+  "typings": "./as-table.d.ts",
+  "version": "1.0.52"
 }
