@@ -1,42 +1,73 @@
 {
-  "name": "get-source",
-  "version": "1.0.27",
+  "_args": [
+    [
+      "get-source@1.0.37",
+      "/Users/Rothacker/Desktop/workspace/trading-bot"
+    ]
+  ],
+  "_from": "get-source@1.0.37",
+  "_id": "get-source@1.0.37",
+  "_inBundle": false,
+  "_integrity": "sha512-Q0YbhnCgvqFe3OTuxkNvWCM+22O4hkx7D0g8em7iFW4vmw/TiX8dhsFIaP6yRkyN9ETHgzU190L3keCg2MADtg==",
+  "_location": "/get-source",
+  "_phantomChildren": {},
+  "_requested": {
+    "type": "version",
+    "registry": true,
+    "raw": "get-source@1.0.37",
+    "name": "get-source",
+    "escapedName": "get-source",
+    "rawSpec": "1.0.37",
+    "saveSpec": null,
+    "fetchSpec": "1.0.37"
+  },
+  "_requiredBy": [
+    "/stacktracey"
+  ],
+  "_resolved": "https://registry.npmjs.org/get-source/-/get-source-1.0.37.tgz",
+  "_spec": "1.0.37",
+  "_where": "/Users/Rothacker/Desktop/workspace/trading-bot",
+  "author": {
+    "name": "Vitaly Gordon",
+    "email": "rocket.mind@gmail.com"
+  },
+  "bugs": {
+    "url": "https://github.com/xpl/get-source/issues"
+  },
+  "dependencies": {
+    "data-uri-to-buffer": "^2.0.0",
+    "source-map": "^0.6.1"
+  },
   "description": "Platform-agnostic source code inspection, with sourcemaps support",
-  "main": "get-source",
-  "scripts": {
-    "test-browser": "mocha test/test.browser --reporter spec",
-    "test": "nyc --reporter=html --reporter=text mocha test/test.path test/test.node --reporter spec",
-    "coveralls": "nyc report --reporter=text-lcov | coveralls"
+  "devDependencies": {
+    "chai": "^3.5.0",
+    "chromedriver": "^2.46.0",
+    "coveralls": "^3.0.3",
+    "istanbul": "^0.4.5",
+    "memory-fs": "^0.3.0",
+    "mocha": "^6.0.2",
+    "nyc": "^13.3.0",
+    "selenium-webdriver": "^3.6.0",
+    "webpack": "^3.12.0"
   },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/xpl/get-source.git"
-  },
+  "homepage": "https://github.com/xpl/get-source",
   "keywords": [
     "sources",
     "sourcemap",
     "read source",
     "cached sources"
   ],
-  "author": "Vitaly Gordon <rocket.mind@gmail.com>",
   "license": "Unlicense",
-  "bugs": {
-    "url": "https://github.com/xpl/get-source/issues"
+  "main": "get-source",
+  "name": "get-source",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/xpl/get-source.git"
   },
-  "homepage": "https://github.com/xpl/get-source",
-  "devDependencies": {
-    "chai": "^3.5.0",
-    "chromedriver": "^2.32.3",
-    "coveralls": "^2.13.1",
-    "istanbul": "^0.4.5",
-    "memory-fs": "^0.3.0",
-    "mocha": "^3.5.3",
-    "nyc": "^11.2.1",
-    "selenium-webdriver": "^2.53.3",
-    "webpack": "^3.6.0"
+  "scripts": {
+    "coveralls": "nyc report --reporter=text-lcov | coveralls",
+    "test": "nyc --reporter=html --reporter=text mocha test/test.path test/test.node --reporter spec",
+    "test-browser": "mocha test/test.browser --reporter spec"
   },
-  "dependencies": {
-    "data-uri-to-buffer": "^2.0.0",
-    "source-map": "^0.6.0"
-  }
+  "version": "1.0.37"
 }
