@@ -1,21 +1,54 @@
 {
-  "name": "stacktracey",
-  "version": "1.2.106",
+  "_args": [
+    [
+      "stacktracey@1.2.117",
+      "/Users/Rothacker/Desktop/workspace/trading-bot"
+    ]
+  ],
+  "_from": "stacktracey@1.2.117",
+  "_id": "stacktracey@1.2.117",
+  "_inBundle": false,
+  "_integrity": "sha512-sj9olGCco8jxMw3fcmhL5er1yuVL2qSMhNfycwx+iIWz6uVpG+bQ2YNpVSlWscGmSskt61gSmUBjZ0Mb5jYBLw==",
+  "_location": "/stacktracey",
+  "_phantomChildren": {},
+  "_requested": {
+    "type": "version",
+    "registry": true,
+    "raw": "stacktracey@1.2.117",
+    "name": "stacktracey",
+    "escapedName": "stacktracey",
+    "rawSpec": "1.2.117",
+    "saveSpec": null,
+    "fetchSpec": "1.2.117"
+  },
+  "_requiredBy": [
+    "/ololog"
+  ],
+  "_resolved": "https://registry.npmjs.org/stacktracey/-/stacktracey-1.2.117.tgz",
+  "_spec": "1.2.117",
+  "_where": "/Users/Rothacker/Desktop/workspace/trading-bot",
+  "author": {
+    "name": "Vitaly Gordon",
+    "email": "rocket.mind@gmail.com"
+  },
+  "bugs": {
+    "url": "https://github.com/xpl/stacktracey/issues"
+  },
+  "dependencies": {
+    "as-table": "^1.0.36",
+    "get-source": "^1.0.34"
+  },
   "description": "Parses call stacks. Reads sources. Clean & filtered output. Sourcemaps. Node & browsers.",
-  "main": "stacktracey",
-  "scripts": {
-    "lint": "eslint stacktracey.js",
-    "lint-test": "eslint stacktracey.js",
-    "babel": "babel stacktracey.js --source-maps inline --out-file ./build/stacktracey.js",
-    "build": "npm run lint && npm run lint-test && npm run babel",
-    "test": "nyc --reporter=html --reporter=text mocha --reporter spec",
-    "autotest": "mocha --reporter spec --watch",
-    "coveralls": "nyc report --reporter=text-lcov | coveralls"
+  "devDependencies": {
+    "babel-cli": "^6.26.0",
+    "chai": "^4.2.0",
+    "coveralls": "^3.0.3",
+    "eslint": "^4.19.1",
+    "istanbul": "^0.4.5",
+    "mocha": "^4.1.0",
+    "nyc": "^14.1.1"
   },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/xpl/stacktracey.git"
-  },
+  "homepage": "https://xpl.github.io/stacktracey",
   "keywords": [
     "stack",
     "stacktrace",
@@ -66,23 +99,21 @@
     "pretty print errors",
     "pretty print stack"
   ],
-  "author": "Vitaly Gordon <rocket.mind@gmail.com>",
   "license": "Unlicense",
-  "bugs": {
-    "url": "https://github.com/xpl/stacktracey/issues"
+  "main": "stacktracey",
+  "name": "stacktracey",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/xpl/stacktracey.git"
   },
-  "homepage": "https://xpl.github.io/stacktracey",
-  "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "chai": "^4.1.2",
-    "coveralls": "^3.0.0",
-    "eslint": "^4.14.0",
-    "istanbul": "^0.4.5",
-    "mocha": "^4.1.0",
-    "nyc": "^11.4.1"
+  "scripts": {
+    "autotest": "mocha --reporter spec --watch",
+    "babel": "babel stacktracey.js --source-maps inline --out-file ./build/stacktracey.js",
+    "build": "npm run lint && npm run lint-test && npm run babel",
+    "coveralls": "nyc report --reporter=text-lcov | coveralls",
+    "lint": "eslint stacktracey.js",
+    "lint-test": "eslint stacktracey.js",
+    "test": "nyc --reporter=html --reporter=text mocha --reporter spec"
   },
-  "dependencies": {
-    "as-table": "^1.0.31",
-    "get-source": "^1.0.24"
-  }
+  "version": "1.2.117"
 }
