@@ -1,21 +1,56 @@
 {
-  "name": "string.ify",
-  "version": "1.0.49",
-  "description": "A cross-platform JavaScript object stringifier / pretty-printer",
-  "main": "build/string.ify",
-  "scripts": {
-    "lint": "eslint string.ify.js",
-    "lint-test": "eslint test.js",
-    "build": "npm run lint && npm run lint-test && npm run babel",
-    "test": "npm run build && env STRING_IFY_TEST_FILE='./build/string.ify' nyc --reporter=html --reporter=text mocha --reporter spec",
-    "autotest": "env STRING_IFY_TEST_FILE='./string.ify' mocha --reporter spec --watch",
-    "coveralls": "nyc report --reporter=text-lcov | coveralls",
-    "babel": "babel string.ify.js --source-maps inline --out-file ./build/string.ify.js"
+  "_args": [
+    [
+      "string.ify@1.0.61",
+      "/Users/Rothacker/Desktop/workspace/trading-bot"
+    ]
+  ],
+  "_from": "string.ify@1.0.61",
+  "_id": "string.ify@1.0.61",
+  "_inBundle": false,
+  "_integrity": "sha512-ROVOYWN8B62bP1LUJIDK6/NooBrN8xIvpV8Ugssbw7Oa4YKJmiOeoEomhDLLRVpYHo0vOBCpEM8RFpxUk9KDIQ==",
+  "_location": "/string.ify",
+  "_phantomChildren": {},
+  "_requested": {
+    "type": "version",
+    "registry": true,
+    "raw": "string.ify@1.0.61",
+    "name": "string.ify",
+    "escapedName": "string.ify",
+    "rawSpec": "1.0.61",
+    "saveSpec": null,
+    "fetchSpec": "1.0.61"
   },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/xpl/string.ify.git"
+  "_requiredBy": [
+    "/ololog"
+  ],
+  "_resolved": "https://registry.npmjs.org/string.ify/-/string.ify-1.0.61.tgz",
+  "_spec": "1.0.61",
+  "_where": "/Users/Rothacker/Desktop/workspace/trading-bot",
+  "author": {
+    "name": "Vitaly Gordon",
+    "email": "rocket.mind@gmail.com"
   },
+  "bugs": {
+    "url": "https://github.com/xpl/string.ify/issues"
+  },
+  "dependencies": {
+    "printable-characters": "^1.0.42",
+    "string.bullet": "^1.0.12"
+  },
+  "description": "A small, simple yet powerful JavaScript object stringifier / pretty-printer",
+  "devDependencies": {
+    "babel-cli": "^6.26.0",
+    "babel-plugin-transform-es2017-object-entries": "0.0.4",
+    "babel-preset-es2016": "^6.24.1",
+    "browserify": "^14.5.0",
+    "coveralls": "^3.0.3",
+    "eslint": "^5.15.1",
+    "istanbul": "^0.4.5",
+    "mocha": "^6.1.4",
+    "nyc": "^14.1.1"
+  },
+  "homepage": "https://github.com/xpl/string.ify",
   "keywords": [
     "stringify",
     "toString",
@@ -24,25 +59,21 @@
     "string",
     "object printer"
   ],
-  "author": "Vitaly Gordon <rocket.mind@gmail.com>",
   "license": "Unlicense",
-  "bugs": {
-    "url": "https://github.com/xpl/string.ify/issues"
+  "main": "build/string.ify",
+  "name": "string.ify",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/xpl/string.ify.git"
   },
-  "homepage": "https://github.com/xpl/string.ify",
-  "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-plugin-transform-es2017-object-entries": "0.0.4",
-    "babel-preset-es2016": "^6.24.1",
-    "browserify": "^14.4.0",
-    "coveralls": "^2.13.1",
-    "eslint": "^4.7.2",
-    "istanbul": "^0.4.5",
-    "mocha": "^3.5.3",
-    "nyc": "^11.2.1"
+  "scripts": {
+    "autotest": "env STRING_IFY_TEST_FILE='./string.ify' mocha --reporter spec --watch",
+    "babel": "babel string.ify.js --source-maps inline --out-file ./build/string.ify.js",
+    "build": "npm run lint && npm run lint-test && npm run babel",
+    "coveralls": "nyc report --reporter=text-lcov | coveralls",
+    "lint": "eslint string.ify.js",
+    "lint-test": "eslint test.js",
+    "test": "npm run build && env STRING_IFY_TEST_FILE='./build/string.ify' nyc --reporter=html --reporter=text mocha --reporter spec"
   },
-  "dependencies": {
-    "printable-characters": "^1.0.28",
-    "string.bullet": "^1.0.11"
-  }
+  "version": "1.0.61"
 }
